app is not defined, can not find mistake, some post methods are not working.
