+++ controllers/comments.js
@@ -1,0 +1,3 @@
+const comments = require("../data/comments");
+
+module.exports = { list, show, create };
