+++ public/index.html
@@ -1,1 +1,14 @@
-<!doctype html><html lang="en"><head><meta charset="utf-8"><meta name="viewport" content="width=device-width,initial-scale=1"><link rel="shortcut icon" href="/favicon.ico"><title>React App</title><link href="/static/css/main.40b4a62f.css" rel="stylesheet"></head><body><div id="root"></div><script type="text/javascript" src="/static/js/main.192f44a5.js"></script></body></html>
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="utf-8" />
+    <meta name="viewport" content="width=device-width,initial-scale=1" />
+    <link rel="shortcut icon" href="/favicon.ico" />
+    <title>React App</title>
+    <link href="/static/css/main.40b4a62f.css" rel="stylesheet" />
+  </head>
+  <body>
+    <div id="root"></div>
+    <script type="text/javascript" src="/static/js/main.192f44a5.js"></script>
+  </body>
+</html>
